requirements.txt
pyserial==3.3
matplotlib==2.0.0

